Change snack price column datatype

+++ db/migrate/20190405161240_create_snacks.rb
@@ -2,7 +2,8 @@ class CreateSnacks < ActiveRecord::Migration[5.1]
   def change
     create_table :snacks do |t|
       t.string :name
-      t.money :price
+      t.float :price
+
       t.references :machine, foreign_key: true
       t.timestamps
     end
